Создание хуков  логика связанная с плейлистом и контролами вынесена в 2 хука
