+++ src/components/audio-player/player.types.ts
@@ -1,13 +1,8 @@
 export interface Track {
-  order?: number;
-  id?: number;
   title: string;
   file: string;
   howl?: Howl | null;
 }
-
-
-
 
 export interface PlayerProps {
   theme?: string;
