+++ src/components/types.ts
@@ -1,0 +1,1 @@
+export type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
