+++ src/custom.d.ts
@@ -2,3 +2,5 @@ declare module "*.svg" {
   const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
   export default content;
 }
+
+
